Update score to use bayesian exploitation with alpha + beta, only use parent nodes for insights now
I realized that if we use the parent nodes in the calculations, we can find ourselves with a "brilliant" local transition and hundreds of terrible macro transitions, and those terrible transitions will have a negative impact on an otherwise brilliant move.

To prevent this, we keep scoring to the local level, and allow the system to self-correct with the Wise Explorer pushing it to equilibrium.