Refined Utility, Certainty, and Score equations + Modifying Wise Explorer to use probability distributions
Utility: Now a simple expected value of the transition/move: (wins - losses)/total

Certainty: How far one standard error is away from 1 (the closer SE is to 0, the more certain we are of our outcome)

Score: Represents the lower-confidence bound of our reading one standard error below the mean. Our score is lower if we have very sparse data, and our score is diluted if ties are high. We don't assign a value to ties, as we want out system to only tie if we aren't certain we'll win (within 1 SE).

Extra:
* Cleaned up wise explorer algorithm, and added documentation to explain new technique.
    The new technique maintains the logic of the original paper, but formalizes it using probability distributions from our table instead of implicit probability distributions from each simulation (which is harder to empirically keep track of)

* Refined debug visualizations to be clearer for multi-step games like chess.

* Added useful documentation to explain how manager.py and other functions work.

* Cleaned up imports, major lint errors, and other tidbits.